move ensure_dtype_not_object from conventions to backends (#9828)
* move ensure_dtype_not_object from conventions to backends
* add whats-new.rst entry
